Scope the summary changes to Azure Pipelines
Only Azure Pipelines can be verified before merging, so the other build
systems go back to rendering exactly what they rendered before.
