+++ src/Cake.GitHubActions.Module/GitHubActionsReportPrinter.cs
@@ -1,4 +1,6 @@
 using System;
+using System.Linq;
+using System.Text;
 using Cake.Common.Build;
 using Cake.Core;
 using Cake.Core.Diagnostics;
@@ -35,7 +37,7 @@ namespace Cake.GitHubActions.Module
             {
                 if (_context.GitHubActions().IsRunningOnGitHubActions)
                 {
-                    _context.GitHubActions().Commands.SetStepSummary(RenderMarkdownReport(report));
+                    WriteToMarkdown(report);
                 }
 
                 WriteToConsole(report);
@@ -62,6 +64,53 @@ namespace Cake.GitHubActions.Module
         public override void WriteStep(string name, Verbosity verbosity)
         {
             // Intentionally left blank
+        }
+
+        private void WriteToMarkdown(CakeReport report)
+        {
+            var includeSkippedReasonColumn = report.Any(r => !string.IsNullOrEmpty(r.SkippedMessage));
+
+            var sb = new StringBuilder();
+            sb.AppendLine(string.Empty);
+
+            if (includeSkippedReasonColumn)
+            {
+                sb.AppendLine("|Task|Duration|Status|Skip Reason|");
+                sb.AppendLine("|----|--------|------|-----------|");
+            }
+            else
+            {
+                sb.AppendLine("|Task|Duration|Status|");
+                sb.AppendLine("|----|--------|------|");
+            }
+
+            foreach (var item in report)
+            {
+                if (ShouldWriteTask(item))
+                {
+                    if (includeSkippedReasonColumn)
+                    {
+                        sb.AppendLine(string.Format("|{0}|{1}|{2}|{3}|", item.TaskName, FormatDuration(item), item.ExecutionStatus.ToReportStatus(), item.SkippedMessage));
+                    }
+                    else
+                    {
+                        sb.AppendLine(string.Format("|{0}|{1}|{2}|", item.TaskName, FormatDuration(item), item.ExecutionStatus.ToReportStatus()));
+                    }
+                }
+            }
+
+            if (includeSkippedReasonColumn)
+            {
+                sb.AppendLine("|||||");
+                sb.AppendLine(string.Format("|**_{0}_**|**_{1}_**|||", "Total:", GetTotalTime(report)));
+            }
+            else
+            {
+                sb.AppendLine("||||");
+                sb.AppendLine(string.Format("|**_{0}_**|**_{1}_**||", "Total:", GetTotalTime(report)));
+            }
+
+            _context.GitHubActions().Commands.SetStepSummary(sb.ToString());
         }
     }
 }
